Move layout support to own namespace.
