+++ llm/turbine_llm/types/gguf_interop/layouts.py
@@ -4,14 +4,9 @@
 # See https://llvm.org/LICENSE.txt for license information.
 # SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-from dataclasses import dataclass
-from typing import Optional
-import warnings
-
-import numpy as np
 import torch
 
-from ..data import (
+from ..tensors import (
     QuantizedTensor,
 )
