+++ llm/turbine_llm/utils/hf_datasets.py
@@ -98,6 +98,20 @@ Dataset(
     ),
 ).alias_to("open_llama_3b_v2_f16_gguf")
 
+Dataset(
+    "SlyEcho/open_llama_3b_v2_q8_0_gguf",
+    (
+        RemoteFile(
+            "gguf", "SlyEcho/open_llama_3b_v2_gguf", "open-llama-3b-v2-q8_0.gguf"
+        ),
+        RemoteFile(
+            "tokenizer_config.json",
+            "openlm-research/open_llama_3b_v2",
+            "tokenizer_config.json",
+            extra_filenames=["tokenizer.model"],
+        ),
+    ),
+).alias_to("open_llama_3b_v2_q8_0_gguf")
 
 ################################################################################
 # Tool entrypoint
